Move IdStorage from GraphCore. Support conjugate LinkId
