+++ src/common/assembly_graph/core/graph.hpp
@@ -9,6 +9,7 @@
 #pragma once
 
 #include "adt/iterator_range.hpp"
+#include "id_storage.hpp"
 #include "observable_graph.hpp"
 #include "coverage.hpp"
 #include "debruijn_data.hpp"
@@ -25,7 +26,7 @@ public:
     typedef base::EdgeId EdgeId;
     typedef base::VertexId VertexId;
     typedef base::VertexIt VertexIt;
-    typedef DataMasterT::LinkId LinkId;
+    typedef omnigraph::impl::LinkId LinkId;
     typedef VertexIt VertexIterator;
     typedef VertexIterator iterator; // for for_each
     typedef const VertexIterator const_iterator; // for for_each
@@ -44,14 +45,15 @@ private:
 
         std::pair<EdgeId, EdgeId> link;
         unsigned overlap;
+        LinkId conjugate_;
     };
 
-    typedef std::vector<Link> LinkStorage;
-    LinkStorage link_storage_;
+    static constexpr unsigned LINK_ID_BIAS = 3;
+    omnigraph::IdStorage<Link> lstorage_;
 
 public:
     DeBruijnGraph(unsigned k)
-            : base(k), coverage_index_(*this), link_storage_{} {}
+            : base(k), coverage_index_(*this), lstorage_(LINK_ID_BIAS) {}
 
     CoverageIndex<DeBruijnGraph>& coverage_index() {
         return coverage_index_;
@@ -76,9 +78,22 @@ public:
         data(v).set_overlap(ovl);
     }
 
+    using base::conjugate;
+
     LinkId add_link(EdgeId e1, EdgeId e2, unsigned ovl) {
-        link_storage_.emplace_back(e1, e2, ovl);
-        return link_storage_.size() - 1;
+        LinkId fwd = lstorage_.create(e1, e2, ovl);
+        if (e1 == this->conjugate(e2)) {
+            lstorage_.at(fwd.int_id()).conjugate_ = fwd;
+            return fwd;
+        }
+        LinkId conj = lstorage_.create(this->conjugate(e2), this->conjugate(e1), ovl);
+        lstorage_.at(fwd.int_id()).conjugate_ = conj;
+        lstorage_.at(conj.int_id()).conjugate_ = fwd;
+        return fwd;
+    }
+
+    LinkId conjugate(LinkId id) const {
+        return lstorage_.at(id.int_id()).conjugate_;
     }
 
     void add_link(VertexId v, LinkId idx) {
@@ -102,7 +117,7 @@ public:
         links.erase(std::remove_if(links.begin(),
                                    links.end(),
                                    [this, &e](const LinkId &link_id) {
-                                     return link_storage_[link_id].link.first == e;
+                                     return lstorage_.at(link_id.int_id()).link.first == e;
                                    }), links.end());
     }
 
@@ -111,7 +126,7 @@ public:
         links.erase(std::remove_if(links.begin(),
                                    links.end(),
                                    [this, &e](const LinkId &link_id) {
-                                       return link_storage_[link_id].link.second == e;
+                                       return lstorage_.at(link_id.int_id()).link.second == e;
                                    }), links.end());
     }
 
@@ -119,8 +134,8 @@ public:
         return data(v).links();
     }
 
-    const auto& link(size_t idx) const {
-        return link_storage_[idx];
+    const auto& link(LinkId idx) const {
+        return lstorage_.at(idx.int_id());
     }
 
     bool is_complex(VertexId v) const {
@@ -141,16 +156,9 @@ public:
         VERIFY_MSG(false, "Link " << in.int_id() << " -> " << out.int_id() << " was not found for vertex " << v.int_id());
     }
 
-    void lreserve(size_t size) { link_storage_.reserve(size); }
+    void lreserve(size_t size) { lstorage_.reserve(size); }
 
-    auto link_begin() { return link_storage_.begin(); }
-    auto link_end() { return link_storage_.end(); }
-    auto link_begin() const { return link_storage_.begin(); }
-    auto link_end() const { return link_storage_.end(); }
-    auto links() { return adt::make_range(link_begin(), link_end()); }
-    auto links() const { return adt::make_range(link_begin(), link_end()); }
-
-    size_t link_size() const {return link_storage_.size(); }
+    size_t link_size() const { return lstorage_.size(); }
 
     using base::AddVertex;
     using base::AddEdge;
