+++ src/common/assembly_graph/core/graph_core.hpp
@@ -8,6 +8,7 @@
 #pragma once
 
 #include "id_distributor.hpp"
+#include "id_storage.hpp"
 #include "utils/verify.hpp"
 #include "utils/logger/logger.hpp"
 #include "utils/stl_utils.hpp"
@@ -76,6 +77,10 @@ struct EdgeId : public Id {
     using Id::Id;
 };
 
+struct LinkId : public Id {
+    using Id::Id;
+};
+
 }
 
 template<class It, class Graph>
@@ -121,6 +126,7 @@ private:
     friend class GraphCore<DataMaster>;
     friend class ConstructionHelper<DataMaster>;
     friend class PairedElementManipulationHelper<EdgeId>;
+    friend class IdStorage<PairedEdge<DataMaster>>;
     //todo unfriend
     friend class PairedVertex<DataMaster>;
     VertexId end_;
@@ -159,6 +165,7 @@ private:
     friend class ConstructionHelper<DataMaster>;
     friend class PairedEdge<DataMaster>;
     friend class PairedElementManipulationHelper<VertexId>;
+    friend class IdStorage<PairedVertex<DataMaster>>;
     template<class It, class Graph>
     friend class conjugate_iterator;
 
@@ -227,121 +234,7 @@ private:
     static constexpr unsigned ID_BIAS = 3;
 
     template<class T>
-    class IdStorage {
-      private:
-        void resize(size_t N) {
-            // INFO("resize to " << N);
-            VERIFY(N > storage_size_);
-            T *new_storage = (T*)malloc(N * sizeof(T));
-            // Note: we cannot iterate over id's here as resize() could
-            // be called during create() and therefore we might end with
-            // uninitialized newly created id as well.
-            for (uint64_t id = bias_; id < storage_size_; ++id) {
-                if (!id_distributor_.occupied(id))
-                    continue;
-
-                T *val = &storage_[id];
-                ::new(new_storage + id) T(std::move(*val));
-            }
-
-            if (storage_)
-                free(storage_);
-            storage_ = new_storage;
-            storage_size_ = N;
-        }
-
-      public:
-        typedef omnigraph::ReclaimingIdDistributor::id_iterator id_iterator;
-        typedef T value_type;
-
-        IdStorage(uint64_t bias = ID_BIAS)
-                : size_(0), bias_(bias), storage_(nullptr), storage_size_(0), id_distributor_(bias) {
-            resize(id_distributor_.size() + bias_);
-        }
-
-        ~IdStorage() {
-            if (storage_) {
-                for (uint64_t id : id_distributor_.ids()) {
-                    // INFO("~Remove " << id << ":" << typeid(T).name());
-
-                    T *val = &storage_[id];
-                    val->~T();
-                }
-
-                free(storage_);
-            }
-        }
-
-        id_iterator id_begin() const { return id_distributor_.begin(); }
-        id_iterator id_end() const { return id_distributor_.end(); }
-        uint64_t max_id() const { return id_distributor_.max_id(); }
-
-        void reserve(size_t sz) {
-            if (storage_size_ >= sz + bias_)
-                return;
-
-            id_distributor_.resize(sz);
-            resize(sz + bias_);
-        }
-
-        // FIXME: Count!
-        size_t size() const noexcept { return size_; }
-
-        bool contains(uint64_t id) const {
-            return id < storage_size_ && id_distributor_.occupied(id);
-        }
-
-        template<typename... ArgTypes>
-        uint64_t create(ArgTypes &&... args) {
-            uint64_t id = id_distributor_.allocate();
-
-            while (storage_size_ < id + 1)
-                resize(storage_size_ * 2 + 1);
-
-            new(storage_ + id) T(std::forward<ArgTypes>(args)...);;
-            size_ += 1;
-
-            // INFO("Create " << id << ":" << typeid(T).name());
-            return id;
-        }
-
-        template<typename... ArgTypes>
-        uint64_t emplace(uint64_t at, ArgTypes &&... args) {
-            // One MUST call reserve before using emplace()
-            VERIFY(!id_distributor_.occupied(at));
-
-            id_distributor_.acquire(at);
-            new(storage_ + at) T(std::forward<ArgTypes>(args)...);;
-            size_.fetch_add(1);
-
-            // INFO("Emplace " << at << ":" << typeid(T).name());
-            return at;
-        }
-
-        void erase(uint64_t id) {
-            T *v = &storage_[id];
-
-            // INFO("Remove " << id << ":" << typeid(T).name());
-            v->~T();
-
-            id_distributor_.release(id);
-            size_ -= 1;
-        }
-
-        T& at(uint64_t id) const noexcept {
-            return storage_[id];
-        }
-
-        uint64_t reserved() const { return id_distributor_.size(); }
-        void clear_state() { id_distributor_.clear_state(); }
-
-      private:
-        std::atomic<size_t> size_;
-        uint64_t bias_;
-        T *storage_;
-        size_t storage_size_;
-        omnigraph::ReclaimingIdDistributor id_distributor_;
-    };
+    using IdStorage = omnigraph::IdStorage<T>;
 
     using VertexStorage = IdStorage<PairedVertex<DataMaster>>;
     VertexStorage vstorage_;
@@ -869,6 +762,13 @@ template<>
 struct hash<omnigraph::impl::EdgeId> {
     size_t operator()(omnigraph::impl::EdgeId e) const noexcept {
         return e.hash();
+    }
+};
+
+template<>
+struct hash<omnigraph::impl::LinkId> {
+    size_t operator()(omnigraph::impl::LinkId l) const noexcept {
+        return l.hash();
     }
 };
