+++ src/common/io/graph/gfa_reader.cpp
@@ -112,8 +112,6 @@ static void HandleLink(Links &links,
         e2 = g.conjugate(e2);
 
     links.emplace_back(e1, e2, record.overlap);
-    if (e1 != g.conjugate(e2))
-        links.emplace_back(g.conjugate(e2), g.conjugate(e1), record.overlap);
 }
 
 static void HandleGapLink(GFAReader::GapLinks &links,
@@ -129,8 +127,6 @@ static void HandleGapLink(GFAReader::GapLinks &links,
         e2 = g.conjugate(e2);
 
     links.emplace_back(e1, e2);
-    if (e1 != g.conjugate(e2))
-        links.emplace_back(g.conjugate(e2), g.conjugate(e1));
 }
 
 
@@ -197,12 +193,17 @@ static std::pair<unsigned, bool> ProcessLinks(DeBruijnGraph &g, const Links &lin
         if (simple) {
             g.set_overlap(v1, ovl);
             g.set_overlap(v2, ovl);
+            g.set_overlap(g.conjugate(v1), ovl);
+            g.set_overlap(g.conjugate(v2), ovl);
         } else {
             LinkId link_idx = g.add_link(e1, e2, ovl);
             g.add_link(v1, link_idx);
+            g.add_link(g.conjugate(v1), g.conjugate(link_idx));
             if (v1 != v2) {
                 g.add_links(v1, g.links(v2));
                 g.clear_links(v2);
+                g.add_links(g.conjugate(v1), g.links(g.conjugate(v2)));
+                g.clear_links(g.conjugate(v2));
             }
         }
